removed not needed parameter

+++ src/parser/parser.c
@@ -48,7 +48,7 @@ static t_command	*parse_tokens(t_shell *shell, t_token *token)
 {
 	t_command	*head;
 	t_command	*cmd;
-	size_t		consumed;
+	int			consumed;
 
 	head = new_command(shell);
 	cmd = head;
